+++ tests/testobjloader.cpp
@@ -255,8 +255,8 @@ void TestObjLoader::OnEvent( Event &evt )
                 double len = std::sqrt(pos[0]*pos[0] + pos[1]*pos[1]);
                 theta += dtheta;
                 glm::vec3 newpos = pos;
-                newpos[0] = len * std::cos(theta);
-                newpos[1] = len * std::sin(theta);
+                newpos[0] = static_cast<float>(len * std::cos(theta));
+                newpos[1] = static_cast<float>(len * std::sin(theta));
                 _camera.SetPosition(newpos);
             }
             break;
@@ -287,7 +287,7 @@ void TestObjLoader::OnEvent( Event &evt )
             const auto& lookAt = _camera.GetLookAt();
             const auto& eye    = _camera.GetPosition();
             glm::vec3 dir = eye - lookAt;
-            float scale = 0.1f * mouseEvt.YOffset();
+            float scale = static_cast<float>(0.1f * mouseEvt.YOffset());
             dir = dir +  dir * scale;
             _camera.SetPosition(lookAt + dir);
             break;
